requirements.txt
asgiref==3.7.2
Django==4.2.0
django-autoslug==1.9.9
django-ckeditor==6.7.0
django-js-asset==2.2.0
django-tinymce==3.6.1
django-versatileimagefield==3.0
pillow==7.2.0
python-magic==0.4.27
sqlparse==0.4.4
tzdata==2023.4